refactor: use providerid type in loginprovider interface and rename file (#719)
