+++ src/components/Login/components/Buttons/types.ts
@@ -1,5 +1,5 @@
 import { ButtonProps } from "@mui/material";
-import { LoginProvider } from "../../../../auth/types/login-provider";
+import { LoginProvider } from "../../../../auth/types/loginProvider";
 import { BaseComponentProps } from "../../../types";
 
 export interface Props extends BaseComponentProps, ButtonProps {
